requirements.txt
anyio==4.6.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
greenlet==3.1.1
h11==0.14.0
pip==24.2
pyasn1==0.6.1
PyMySQL==1.1.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.35
annotated-types==0.7.0
bcrypt==4.2.0
cffi==1.17.1
cryptography==43.0.1
fastapi==0.115.2
idna==3.10
jwt==1.3.1
passlib==1.7.4
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-multipart==0.0.12
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
